// src/app/model/ClaimSheetReport.ts
export class ClaimSheetReport {
  IRFID: number;
  IRFType: string;
  IRFSubType: string;
  InvoiceNumber: string;
  CreditNoteAmount: string;
  InvoiceDate: Date;
  ClientName: string;
  ClientType: string;
  InvoiceAmountBeforeTax: number;
  InvoiceAmountAfterTax: number;
  CollectedAmount: number;
  OutStandingAmount: number;
  GrossProfit: number;
  EntityType: string;
  EntityId: number;
  EntityName: string;
  SharingType: string;
  Value: number;
  ComputedValue: number;
  ConsultantName: string;
  Id: number;
  IRFREVSHAREID: number;
  INVOICEID: number;
  BusinessUnit: string;
}