// src/app/model/ClaimSheet.ts
export class ClaimSheet{
  IRFID: number;
  IRFType: string;
  InvoiceNumber: string;
  InvoiceDate: Date;
  CreditNoteAmount: number;
  ClientName: string;
  ClientType: string;
  InvoiceAmountBeforeTax: number;
  InvoiceAmountAfterTax: number;
  CollectedAmount: number;
  OutStandingAmount: number;
  GrossProfit: number;
  EntityType: string;
  EntityId: number;
  EntityName: string;
  SharingType: string;
  Value: number;
  ComputedValue: number;
  ConsultantName: string;
  Id: number;
  RequestedBy: string;
  RequestedAmount: number;
  RequesterRemarks: string;
  ApprovedAmount: number;
  ApprovedOn: Date;
  ApprovedBy: string;
  ApproverRemarks: string;
  Status: number;
  CreatedOn: Date;
  CreatedBy: string;
  LastUpdatedOn: Date;
  LastUpdatedBy: string;
  ClaimSheetRequestId: number;
  IRFRevenueSharingId: number;
  InvoiceId: number;
}