// src/app/model/ClaimSheetDetail.ts
export class ClaimSheetDetail{
  IRFID: number;
  IRFType: string;
  IRFSubType: string;
  InvoiceNumber: string;
  InvoiceDate: Date;
  CreditNoteAmount: number;
  ClientName: string;
  ClientType: string;
  InvoiceAmountBeforeTax: number;
  InvoiceAmountAfterTax: number;
  CollectedAmount: number;
  OutStandingAmount: number;
  GrossProfit: number;
  EntityType: string;
  EntityId: number;
  EntityName: string;
  SharingType: string;
  Value: number;
  ComputedValue: number;
  ConsultantName: string;
  Id: number;
  RequestedBy: string;
  RequestedAmount: number;
  RequesterRemarks: string;
  ApprovedAmount: number;
  ApprovedOn: Date;
  ApprovedBy: string;
  ApproverRemarks: string;
  GrossAmount: number;
  Status: string;
  StatusText: string;
  CreatedOn: Date;
  CreatedOnString: string;
  CreatedBy: string;
  LastUpdatedOn: Date;
  LastUpdatedBy: string;
  ClaimSheetRequestId: number;
  IRFRevenueSharingId: number;
  InvoiceId: number;
  NetClaimAmount: number;
  LicenseCost: number;
  Advances: number;
  FinanceCost: number;
  PortalFee: number;
  MiscellaneousAdjustmentAmount: number;
  SupportingDocumentId: number;
  FileName: string;
  IRFList: string;
  IRFArray: Array<ClaimSheetDetail>;
  IRFREVSHAREID: number;
  INVOICEID: number;
  InvoiceAmount: number;
}